fix: remove all comment nodes when exporting
